requirements.txt
aiofiles==0.7.0
aiogram==2.15
aiohttp==3.7.4.post0
altgraph==0.17
antigravity==0.1
anyio==3.3.0
asgiref==3.4.1
async-generator==1.10
async-lru==1.0.2
async-timeout==3.0.1
attrs==21.2.0
auto-py-to-exe==2.10.0
b0mb3r==0.1.0
Babel==2.9.1
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
config==0.5.1
cryptography==35.0.0
db0mb3r==3.1.4.17
dearpygui==0.8.64
discord-interactions==0.3.0
discord-interactions.py @ git+https://github.com/LiBa001/discord-interactions.py@7f19e1b33dfdf2fd9d0a95a164fba8820a031bba
discord-webhook==0.14.0
discord.py==1.7.3
Eel==0.12.4
fake-useragent==0.1.11
fastapi==0.68.1
Flask==2.0.1
future==0.18.2
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
keyboard==0.13.5
loguru==0.5.3
MarkupSafe==2.0.1
mcpi-e==0.3.2006.2
MouseInfo==0.1.3
multidict==5.1.0
numpy==1.21.2
opencv-python==4.5.3.56
outcome==1.1.0
pefile==2021.5.24
phonenumbers==8.12.31
Pillow==8.3.2
prettytable==2.2.0
py3dns==3.2.1
pyaes==1.6.1
PyAutoGUI==0.9.53
pybase64==1.2.0
pycparser==2.20
pydantic==1.8.2
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
PyRect==0.1.4
Pyrogram==1.2.9
PyScreeze==0.1.28
PySide2==5.15.2
PySimpleGUI==4.49.0
PySocks==1.7.1
python-dotenv==0.19.0
pytweening==1.0.4
pytz==2021.3
pywebio==1.4.0
pywin32-ctypes==0.2.0
pyxel==1.4.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
random-user-agent==1.0.1
requests==2.26.0
rfc3986==1.5.0
rich==10.9.0
selenium==4.0.0
shiboken2==5.15.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.14.2
termcolor==1.1.0
tornado==6.1
tqdm==4.62.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
uvicorn==0.15.0
validate-email==1.3
vk-api==11.9.4
wcwidth==0.2.5
Werkzeug==2.0.1
whichcraft==0.6.1
win32-setctime==1.0.3
wsproto==1.0.0
yarl==1.6.3
